requirements.txt
# Instagram Engagement Prediction - Requirements
# Minimal dependencies for MVP approach

# Core Data Science
pandas==2.1.0
numpy==1.24.3
scipy==1.11.2

# Machine Learning
scikit-learn==1.3.0
xgboost==2.0.0

# Indonesian NLP (Phase 2)
Sastrawi==1.2.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1

# Model Interpretability (Phase 2)
shap==0.42.1

# Web Application
streamlit==1.26.0

# Utilities
pyyaml==6.0.1
joblib==1.3.2
tqdm==4.66.1
python-dotenv==1.0.0

# Development
jupyter==1.0.0
ipykernel==6.25.0
black==23.7.0
flake8==6.1.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
